+++ src/main/kotlin/icu/windea/pls/lang/codeInsight/markers/script/ParadoxDefineVariableLineMarkerProvider.kt
@@ -15,54 +15,52 @@ import icu.windea.pls.core.util.values.anonymous
 import icu.windea.pls.core.util.values.or
 import icu.windea.pls.lang.actions.PlsActions
 import icu.windea.pls.lang.codeInsight.markers.ParadoxRelatedItemLineMarkerProvider
-import icu.windea.pls.lang.definitionInfo
+import icu.windea.pls.lang.defineVariableInfo
+import icu.windea.pls.lang.search.ParadoxDefineNamespaceSearch
+import icu.windea.pls.lang.search.ParadoxDefineVariableSearch
 import icu.windea.pls.lang.search.ParadoxDefinitionSearch
 import icu.windea.pls.lang.search.selector.contextSensitive
 import icu.windea.pls.lang.search.selector.selector
-import icu.windea.pls.model.ParadoxDefinitionSource
 import icu.windea.pls.model.constants.PlsStrings
 import icu.windea.pls.script.psi.ParadoxScriptProperty
 
 /**
- * 提供定义（definition）的装订线图标。
+ * 提供定义变量（defineVariable）的装订线图标。
  */
-class ParadoxDefinitionLineMarkerProvider : ParadoxRelatedItemLineMarkerProvider() {
-    override fun getName() = PlsBundle.message("script.gutterIcon.definition")
+class ParadoxDefineVariableLineMarkerProvider : ParadoxRelatedItemLineMarkerProvider() {
+    override fun getName() = PlsBundle.message("script.gutterIcon.defineVariable")
 
-    override fun getIcon() = PlsIcons.Gutter.Definition
+    override fun getIcon() = PlsIcons.Gutter.DefineVariable
 
-    override fun getGroup() = PlsBundle.message("script.gutterIcon.definition.group")
+    override fun getGroup() = PlsBundle.message("script.gutterIcon.defineVariable.group")
 
     override fun collectNavigationMarkers(element: PsiElement, result: MutableCollection<in RelatedItemLineMarkerInfo<*>>) {
-        // 何时显示装订线图标：element 是 definition
+        // 何时显示装订线图标：element 是 defineVariable
         if (element !is ParadoxScriptProperty) return
         val locationElement = element.propertyKey.idElement ?: return
-        val definitionInfo = element.definitionInfo ?: return
-
-        // 忽略内联或注入的定义
-        if (definitionInfo.source == ParadoxDefinitionSource.Inline || definitionInfo.source == ParadoxDefinitionSource.Injection) return
+        val defineVariableInfo = element.defineVariableInfo ?: return
 
         ProgressManager.checkCanceled()
-        val icon = PlsIcons.Gutter.Definition
-        val prefix = PlsStrings.definitionPrefix
-        val tooltip = "$prefix <b>${definitionInfo.name.escapeXml().or.anonymous()}</b>: ${definitionInfo.typeText}"
+        val icon = PlsIcons.Gutter.DefineVariable
+        val prefix = PlsStrings.defineVariablePrefix
+        val tooltip = "$prefix <b>${defineVariableInfo.namespace.escapeXml()}.${defineVariableInfo.variable.escapeXml()}</b>"
         val targets by lazy {
             val project = element.project
-            val selector = selector(project, element).definition().contextSensitive()
-            val targets0 = ParadoxDefinitionSearch.searchElement(definitionInfo.name, definitionInfo.type, selector).findAll()
+            val selector = selector(project, element).define().contextSensitive()
+            val targets0 = ParadoxDefineVariableSearch.search(defineVariableInfo.namespace, defineVariableInfo.variable, selector).findAll()
             targets0.optimized()
         }
         val lineMarkerInfo = NavigationGutterIconBuilderFacade.createForPsi(icon) { createGotoRelatedItem(targets) }
             .setTooltipText(tooltip)
-            .setPopupTitle(PlsBundle.message("script.gutterIcon.definition.title"))
+            .setPopupTitle(PlsBundle.message("script.gutterIcon.defineVariable.title"))
             .setTargets { targets }
             .setAlignment(GutterIconRenderer.Alignment.LEFT)
-            .setNamer { PlsBundle.message("script.gutterIcon.definition") }
+            .setNamer { PlsBundle.message("script.gutterIcon.defineVariable") }
             .createLineMarkerInfo(locationElement)
         result.add(lineMarkerInfo)
 
         // 绑定导航动作 & 在单独的分组中显示对应的意向动作
-        val actionText = PlsBundle.message("script.gutterIcon.definition.action")
-        NavigateAction.setNavigateAction(lineMarkerInfo, actionText, PlsActions.GotoDefinitions)
+        val actionText = PlsBundle.message("script.gutterIcon.defineVariable.action")
+        NavigateAction.setNavigateAction(lineMarkerInfo, actionText, PlsActions.GotoDefineVariables)
     }
 }