+++ src/main/kotlin/icu/windea/pls/lang/codeInsight/markers/script/ParadoxDefinitionInjectionsLineMarkerProvider.kt
@@ -61,10 +61,7 @@ class ParadoxDefinitionInjectionsLineMarkerProvider : ParadoxRelatedItemLineMark
         result.add(lineMarkerInfo)
 
         // 绑定导航动作 & 在单独的分组中显示对应的意向动作
-        NavigateAction.setNavigateAction(
-            lineMarkerInfo,
-            PlsBundle.message("script.gutterIcon.definitionInjections.action"),
-            PlsActions.GotoDefinitionInjections
-        )
+        val actionText = PlsBundle.message("script.gutterIcon.definitionInjections.action")
+        NavigateAction.setNavigateAction(lineMarkerInfo, actionText, PlsActions.GotoDefinitionInjections)
     }
 }