+++ src/main/kotlin/icu/windea/pls/lang/codeInsight/markers/script/ParadoxDefinitionRelatedImagesLineMarkerProvider.kt
@@ -72,10 +72,7 @@ class ParadoxDefinitionRelatedImagesLineMarkerProvider : ParadoxRelatedItemLineM
         result.add(lineMarkerInfo)
 
         // 绑定导航动作 & 在单独的分组中显示对应的意向动作
-        NavigateAction.setNavigateAction(
-            lineMarkerInfo,
-            PlsBundle.message("script.gutterIcon.definitionRelatedImages.action"),
-            PlsActions.GotoRelatedImages
-        )
+        val actionText = PlsBundle.message("script.gutterIcon.definitionRelatedImages.action")
+        NavigateAction.setNavigateAction(lineMarkerInfo, actionText, PlsActions.GotoRelatedImages)
     }
 }