+++ src/main/kotlin/icu/windea/pls/lang/codeInsight/markers/script/ParadoxInlineScriptsLineMarkerProvider.kt
@@ -54,10 +54,7 @@ class ParadoxInlineScriptsLineMarkerProvider : ParadoxRelatedItemLineMarkerProvi
         result.add(lineMarkerInfo)
 
         // 绑定导航动作 & 在单独的分组中显示对应的意向动作
-        NavigateAction.setNavigateAction(
-            lineMarkerInfo,
-            PlsBundle.message("script.gutterIcon.inlineScripts.action"),
-            PlsActions.GotoInlineScripts
-        )
+        val actionText = PlsBundle.message("script.gutterIcon.inlineScripts.action")
+        NavigateAction.setNavigateAction(lineMarkerInfo, actionText, PlsActions.GotoInlineScripts)
     }
 }